requirements.txt
openai
langchain
faiss-cpu
chromadb
python-dotenv
langchain-openai
langchain-community
